refactor[#12]; managed_action_log 삭제 로직 resizer로 이전
